refactor files locatios
