+++ src/category/controllers/__test__/category.controller.spec.ts
@@ -1,12 +1,12 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { CategoryController } from './category.controller';
-import { CategoryService } from './category.service';
-import { CategoryModule } from './category.module';
+import { CategoryController } from '../category.controller';
+import { CategoryService } from '../../services/category.service';
+import { CategoryModule } from '../../category.module';
 import { CATEGORY_REPOSITORY } from 'src/shared/utils/tokens';
 import { WinstonLoggerService } from 'src/shared/infrastructure/logger/winston-logger.service';
 import { TYPES } from 'src/shared/utils/types';
 import { ConfigService } from '@nestjs/config';
-import { CategoryEntity } from './entities/category.entity';
+import { CategoryEntity } from '../../entities/category.entity';
 import path from 'node:path';
 import { readFile } from 'node:fs/promises';
 
@@ -84,4 +84,3 @@ describe('CategoryController', () => {
     expect(controller).toBeDefined();
   });
 });
-
