+++ src/category/services/category.service.ts
@@ -1,13 +1,13 @@
 import { Inject, Injectable, InternalServerErrorException } from '@nestjs/common';
-import { CreateCategoryDto } from './dto/create-category.dto';
-import { UpdateCategoryDto } from './dto/update-category.dto';
-import { CategoryRepository } from './repositories/category.repository';
+import { CreateCategoryDto } from '../dto/create-category.dto';
+import { UpdateCategoryDto } from '../dto/update-category.dto';
+import { CategoryRepository } from '../repositories/category.repository';
 import { CATEGORY_REPOSITORY } from 'src/shared/utils/tokens';
-import { CategoryEntity } from './entities/category.entity';
+import { CategoryEntity } from '../entities/category.entity';
 import { LoggerService } from 'src/shared/infrastructure/logger/logger.service';
 import { TYPES } from 'src/shared/utils/types';
-import { SubcategoryLevel2 } from './dto/subcategoryLevel2.dto';
-import { Category } from './dto/category.dto';
+import { SubcategoryLevel2 } from '../dto/subcategoryLevel2.dto';
+import { Category } from '../dto/category.dto';
 
 @Injectable()
 export class CategoryService {
