+++ src/coupon/services/__test__/coupon.service.spec.ts
@@ -1,31 +1,22 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { CouponController } from './coupon.controller';
-import { CouponService } from './coupon.service';
-import { CouponEntity } from './entities/coupon.entity';
+import { CouponService } from '../coupon.service';
+import { CouponEntity } from '../../entities/coupon.entity';
 import { COUPON_REPOSITORY } from 'src/shared/utils/tokens';
 import { CACHE_MANAGER } from '@nestjs/cache-manager';
-import { CouponModule } from './coupon.module';
 import { WinstonLoggerService } from 'src/shared/infrastructure/logger/winston-logger.service';
 import { TYPES } from 'src/shared/utils/types';
 import { ConfigService } from '@nestjs/config';
-import fs, { readFile } from 'node:fs/promises';
+import { CouponModule } from '../../coupon.module';
 import path from 'node:path';
+import { readFile } from 'node:fs/promises';
 
-describe('CouponController', () => {
-  let controller: CouponController;
+describe('CouponService', () => {
   let service: CouponService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [CouponModule],
-      controllers: [CouponController],
       providers: [CouponService,
-        {
-          provide: COUPON_REPOSITORY,
-          useValue: {
-            find: jest.fn(),
-          },
-        },
         {
           provide: CACHE_MANAGER,
           useValue: {
@@ -49,6 +40,13 @@ describe('CouponController', () => {
           },
         },
         {
+          provide: COUPON_REPOSITORY,
+          useValue: {
+            get: jest.fn(),
+            set: jest.fn(),
+          }
+        },
+        {
           provide: ConfigService,
           useValue: {
             get: jest.fn(),
@@ -57,10 +55,10 @@ describe('CouponController', () => {
         }],
     }).compile();
 
-    controller = module.get<CouponController>(CouponController);
     service = module.get<CouponService>(CouponService);
   });
 
+  
   describe('findNotExpired', () => {
     it('should return an array of coupons', async () => {
       const relativePath = 'test/data/coupons-ok.json';
@@ -70,16 +68,15 @@ describe('CouponController', () => {
 
       jest.spyOn(service, 'findNotExpired').mockImplementation(async () => result);
 
-      const found = await controller.findNotExpired();
+      const found = await service.findNotExpired();
       expect(found.length).toBe(result.length);
-
+       
       expect(found[0].expiresAt).toBe(result[0].expiresAt);
       expect(found[0].description).toBe(result[0].description);
     });
   });
-
+  
   it('should be defined', () => {
-    expect(controller).toBeDefined();
+    expect(service).toBeDefined();
   });
 });
-
